+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 # v3.1.0
+* Changed: Set Elasticsearch scroll api query to have a size that is the same as page size
 * Changed: Properly throw NoSuchElementException from Iterables when next is called with no more elements
 * Added: Accumulo in table storage of streaming property value data
 * Added: Functions to delete extended data columns